[#101] REFACTOR: 패키지 경로 정리, 누락된 어노테이션 추가
