+++ backend/src/main/java/com/team5/secondhand/api/member/dto/request/MemberJoin.java
@@ -3,12 +3,15 @@ package com.team5.secondhand.api.member.dto.request;
 import com.team5.secondhand.api.member.domain.Member;
 import com.team5.secondhand.api.member.domain.Oauth;
 import lombok.Data;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
 
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
 import java.util.List;
 
-@Data
+@Getter
+@RequiredArgsConstructor
 public class MemberJoin {
 
     @NotNull
