+++ backend/src/main/java/com/team5/secondhand/api/region/repository/RegionRepository.java
@@ -1,6 +1,6 @@
 package com.team5.secondhand.api.region.repository;
 
-import com.team5.secondhand.api.model.Region;
+import com.team5.secondhand.api.region.domain.Region;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
