+++ backend/src/main/java/com/team5/secondhand/api/region/service/GetValidRegionsUsecase.java
@@ -1,6 +1,6 @@
 package com.team5.secondhand.api.region.service;
 
-import com.team5.secondhand.api.model.Region;
+import com.team5.secondhand.api.region.domain.Region;
 import com.team5.secondhand.api.region.exception.NotValidRegionException;
 
 import java.util.List;
