+++ backend/src/main/java/com/team5/secondhand/global/exception/GlobalExceptionHandler.java
@@ -6,12 +6,15 @@ import com.team5.secondhand.global.aws.exception.TooLargeImageException;
 import com.team5.secondhand.global.dto.ErrorResponse;
 import com.team5.secondhand.global.dto.ErrorResponseWithBody;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.core.Ordered;
+import org.springframework.core.annotation.Order;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
 @Slf4j
+@Order(Ordered.LOWEST_PRECEDENCE)
 @RestControllerAdvice
 public class GlobalExceptionHandler {
